BWS-71: + make pagination components
