--- frontend/src/common/constants/workers.constants.ts
@@ -1,11 +1,0 @@
-const WORKERS_PER_PAGE = 5;
-const WORKER_BACK_PAGE = 5;
-const WORKER_NEXT_PAGE = 5;
-const WORKER_CURRENT_PAGE = 1;
-
-export {
-  WORKERS_PER_PAGE,
-  WORKER_BACK_PAGE,
-  WORKER_NEXT_PAGE,
-  WORKER_CURRENT_PAGE,
-};
