+++ frontend/src/common/enums/enums.ts
@@ -46,3 +46,4 @@ export {
   NotificationMessage,
   NotificationType,
 } from './notification/notification';
+export { Page } from './worker/workers.enum';
