+++ frontend/src/common/enums/worker/workers.enum.ts
@@ -1,0 +1,8 @@
+enum Page {
+  PER_PAGE = 5,
+  BACK_PAGE = 5,
+  NEXT_PAGE = 5,
+  CURRENT_PAGE = 1,
+}
+
+export { Page };
