+++ frontend/src/components/eam/eam.tsx
@@ -1,34 +1,17 @@
 import { FC } from 'react';
 import { Button, IconButton } from 'components/common/common';
 import { AppRoute, IconName } from 'common/enums/enums';
-import {
-  useAppDispatch,
-  useAppSelector,
-  useEffect,
-  useState,
-} from 'hooks/hooks';
+import { useAppDispatch, useAppSelector, useEffect } from 'hooks/hooks';
 import { eam as eamActions } from 'store/actions';
 import { GroupsTable, WorkersTable, Tenant } from './components/components';
 import styles from './styles.module.scss';
-import {
-  WORKERS_PER_PAGE,
-  WORKER_BACK_PAGE,
-  WORKER_NEXT_PAGE,
-  WORKER_CURRENT_PAGE,
-} from 'common/constants/workers.constants';
 
 const EAM: FC = () => {
   const dispatch = useAppDispatch();
-  const [currentPage, setCurrentPage] = useState(1);
-  const [from, setFrom] = useState(0);
 
-  const { id: tenantId, countItems: countItems } = useAppSelector(
-    ({ app, eam }) => ({
-      id: app.tenant?.id,
-      countItems: eam.countItems,
-    }),
-  );
-  const allPage = Math.ceil(countItems / WORKERS_PER_PAGE);
+  const { id: tenantId } = useAppSelector(({ app }) => ({
+    id: app.tenant?.id,
+  }));
 
   useEffect(() => {
     if (!tenantId) {
@@ -75,35 +58,6 @@ const EAM: FC = () => {
     );
   };
 
-  const handleLoadWorkers = (from: number): void => {
-    dispatch(
-      eamActions.loadWorkers({
-        tenantId: tenantId as string,
-        from: from,
-        count: 5,
-      }),
-    );
-  };
-
-  const handleBackPage = (): void => {
-    const backPage = currentPage - WORKER_CURRENT_PAGE;
-    if (backPage !== 0) {
-      setCurrentPage(backPage);
-      const backFrom = from - WORKER_BACK_PAGE;
-      setFrom(backFrom);
-      handleLoadWorkers(from - WORKER_BACK_PAGE);
-    }
-  };
-  const handleNextPage = (): void => {
-    const nextPage = currentPage + WORKER_CURRENT_PAGE;
-    if (nextPage <= allPage) {
-      setCurrentPage(nextPage);
-      const nextForm = from + WORKER_NEXT_PAGE;
-      setFrom(nextForm);
-      handleLoadWorkers(from + WORKER_NEXT_PAGE);
-    }
-  };
-
   return (
     <div className={styles.wrapper}>
       <h2 className={styles.title}>
@@ -112,29 +66,7 @@ const EAM: FC = () => {
       </h2>
       <Tenant />
       <div className={styles.tableWrapper}>
-        <WorkersTable
-          footer={
-            <div className={styles.pagination}>
-              <div className={styles.countItems}>
-                <div className={styles.count}>{countItems + 'results'}</div>
-              </div>
-              <div className={styles.currentPage}>
-                <IconButton
-                  onClick={handleBackPage}
-                  icon={IconName.ARROW_LEFT}
-                  label="ArrowRight"
-                />
-                <div>{currentPage}</div>
-                <div className={styles.count}>of {allPage}</div>
-                <IconButton
-                  onClick={handleNextPage}
-                  icon={IconName.ARROW_RIGHT}
-                  label="ArrowRight"
-                />
-              </div>
-            </div>
-          }
-        >
+        <WorkersTable>
           <div className={styles.buttonsBlock}>
             <IconButton
               onClick={handleWorkersReload}
